% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateHyperbolic.R
\name{calculateHyperbolic}
\alias{calculateHyperbolic}
\title{A function that returns the hyperbolic value of the given input.}
\usage{
calculateHyperbolic(data)
}
\arguments{
\item{data}{The input data}
}
\value{
The hyperboic value of the given data
}
\description{
A function that returns the hyperbolic value of the given input.
}

