% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSigmoid.R
\name{calculateSigmoid}
\alias{calculateSigmoid}
\title{A function that returns the sigmoid value of the given input.}
\usage{
calculateSigmoid(data)
}
\arguments{
\item{data}{The input data}
}
\value{
The sigmoid value of the given data
}
\description{
A function that returns the sigmoid value of the given input.
}

