% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSSRMatrix.nn.R
\name{calculateSSRMatrix.nn}
\alias{calculateSSRMatrix.nn}
\title{A function that calculates and returns the ssr of all possible combinations of segments in a matrix form.
The index of row and column represents the starting and ending point of the segment.}
\usage{
calculateSSRMatrix.nn(data)
}
\arguments{
\item{data}{The data that needs to be approximated}
}
\value{
SSRMatrix The matrix that contains the sums of square residuals (SSR) of all possible segments for the given dataset
}
\description{
A function that calculates and returns the ssr of all possible combinations of segments in a matrix form.
The index of row and column represents the starting and ending point of the segment.
}

