% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateWeightedValues.R
\name{calculateWeightedValues}
\alias{calculateWeightedValues}
\title{A function that takes an input matrix and a weight matrix and return a weighted input matrix.}
\usage{
calculateWeightedValues(Data, WeightMatrix)
}
\arguments{
\item{Data}{The input data}

\item{WeightMatrix}{The weight values}
}
\value{
The weighted Input
}
\description{
A function that takes an input matrix and a weight matrix and return a weighted input matrix.
}

