% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findbestpwl.R
\name{findbestpwl}
\alias{findbestpwl}
\title{A function that is used to find the best set of pwl in that will return a mse% difference of less than specified error.}
\usage{
findbestpwl(Inputdata, ANNOutput, ActualOutput, hiddenW, outputW,
  actfun = "sigmoid", allpwlfull, error)
}
\arguments{
\item{Inputdata}{The input data used to trained the ANN model.}

\item{ANNOutput}{The output predicted by the ANN model.}

\item{ActualOutput}{The target output used to train the ANN model.}

\item{hiddenW}{The weights of the hidden neuron in matrix form, columns are the hidden nodes, rows are each of the weights to the hidden node. Only one hidden layer is supported.}

\item{outputW}{The output weights of the output layer. The output activation function is assumed to be linear.}

\item{actfun}{The activation function of the hidden layer. It can be either "sigmoid" or "hyperbolic". Default function is "sigmoid".}

\item{allpwlfull}{The full set of pwl equations for all hidden nodes.}

\item{Error}{The accepted error rate in terms of percentage difference in MSE with the ANN model. The default value is 20%.}
}
\value{
The set of pwl equations that satisfies the given error
}
\description{
A function that is used to find the best set of pwl in that will return a mse% difference of less than specified error.
}

