% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findBP.R
\name{findBP}
\alias{findBP}
\title{A function that finds the breakpoints (BP) with the minimum ssr, given the number of desired BP, by searching through the given ssr matrix.}
\usage{
findBP(SSRMatrix, noOfBP, l, index, dataSize)
}
\arguments{
\item{SSRMatrix}{The SSR matrix}

\item{noOfBP}{The desired number of breakpoint}

\item{l}{The minimum distance between breakpoints}

\item{index}{The index of the to start the search from}

\item{dataSize}{The data size for the data that is being approximated}
}
\value{
The list of breakpoints and the SSR
}
\description{
A function that finds the breakpoints (BP) with the minimum ssr, given the number of desired BP, by searching through the given ssr matrix.
}

