% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findIntercept.r
\name{findIntercept}
\alias{findIntercept}
\title{A function that determines the x-intercept of two given lines.}
\usage{
findIntercept(intercept1, gradient1, intercept2, gradient2)
}
\arguments{
\item{intercept1}{The intercept of line 1}

\item{gradient1}{The gradient of line 1}

\item{intercept2}{The intercept of line 2}

\item{gradient2}{The gradient of line 2}
}
\value{
The x-intercept of the two given lines
}
\description{
A function that determines the x-intercept of two given lines.
}

