% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findoneBP.nn.R
\name{findoneBP.nn}
\alias{findoneBP.nn}
\title{A function that is used to locate one breakpoint only with brutte force search, without creating the SSR matrix.}
\usage{
findoneBP.nn(data, l)
}
\arguments{
\item{data}{The data that needs to be approximated}

\item{l}{The minimum length of a segment}
}
\value{
The approximated piecewise linear equations
}
\description{
A function that is used to locate one breakpoint only with brutte force search, without creating the SSR matrix.
}

