% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getequations.nn.R
\name{getequations.nn}
\alias{getequations.nn}
\title{A function that gets the equations when given the breakpoints.}
\usage{
getequations.nn(data, BPs)
}
\arguments{
\item{data}{The data being approximated}

\item{BPs}{The list of breakpoints}
}
\value{
Coefficients of the equations, fitted values, and residuals.
}
\description{
A function that gets the equations when given the breakpoints.
}

