% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfullpwl.r
\name{getfullpwl}
\alias{getfullpwl}
\title{A function that finds the "left hand side" equation for each node. In order the maintain continuity of the curve, the intercepts of the lines become the breakpoints.}
\usage{
getfullpwl(pwl, data, constraint)
}
\arguments{
\item{pwl}{The right hand side pwl equations}

\item{data}{The data that is being approximated}

\item{constraint}{The constraint sets by the data, e.g. sigmoid function must pass through y = 0.5 at x = 0}
}
\value{
The full set of pwl equations (letf and right side) of the data
}
\description{
A function that finds the "left hand side" equation for each node. In order the maintain continuity of the curve, the intercepts of the lines become the breakpoints.
}

