% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ismidsection.R
\name{ismidsection}
\alias{ismidsection}
\title{A function that checks if the given data is a "mid" section, i.e. if the data set crosses the x=0
This is used in getequataion.nn and calculateSSRmatrix. If the data is a mid section, then it has to pass through the origin, i.e. lm(y~x+0)}
\usage{
ismidsection(data)
}
\arguments{
\item{data}{The data that needs to be checked}
}
\value{
TRUE if it is a mid section, otherwise FALSE
}
\description{
A function that checks if the given data is a "mid" section, i.e. if the data set crosses the x=0
This is used in getequataion.nn and calculateSSRmatrix. If the data is a mid section, then it has to pass through the origin, i.e. lm(y~x+0)
}

