% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicts.pwlnn.R
\name{predicts.pwlnn}
\alias{predicts.pwlnn}
\title{A function that predicts the pwlnn output.}
\usage{
predicts.pwlnn(Inputdata, ANNOutput, ActualOutput, hiddenW, outputW, allpwlfull,
  actfun = "sigmoid")
}
\arguments{
\item{Inputdata}{The input data used to trained the ANN model.}

\item{ANNOutput}{The output predicted by the ANN model.}

\item{ActualOutput}{The target output used to train the ANN model.}

\item{hiddenW}{The weights of the hidden neuron in matrix form, columns are the hidden nodes, rows are each of the weights to the hidden node. Only one hidden layer is supported.}

\item{outputW}{The output weights of the output layer. The output activation function is assumed to be linear.}

\item{allpwlfull}{The full set of pwl equations for all hidden nodes.}

\item{actfun}{The activation function of the hidden layer. It can be either "sigmoid" or "hyperbolic". Default function is "sigmoid".}
}
\value{
The predicted output given by the pwl-ann model.
}
\description{
A function that predicts the pwlnn output.
}

