% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwl.nn.R
\name{pwl.nn}
\alias{pwl.nn}
\title{A function that finds the piecewise linear approximation function for the given data and number of breakpoints (bp). This pwl.nn() function is modified to include the constraints that are given by the activation functions.}
\usage{
pwl.nn(data, noOfBP = NULL, l, maxBP = NULL, ...)
}
\arguments{
\item{data}{The data that user wants to approximate}

\item{noOfBP}{The desired number of breakpoints}

\item{l}{The minumum distance between breakpoints}

\item{maxBP}{The user can either provide number of desired breakpoints, or the desired maximum number of breakpoints. When maxBP is given, the function returns all the PWL equations from 1 to maxBP.}
}
\value{
a class "pwl" which includes the coefficients of the equations, loaction of the breakpoints, fitted values, residuals and mse.
}
\description{
A function that finds the piecewise linear approximation function for the given data and number of breakpoints (bp). This pwl.nn() function is modified to include the constraints that are given by the activation functions.
}

