% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwlnn.R
\name{pwlnn}
\alias{pwlnn}
\title{pwlnn - piecewise linear neural network.
A function that approximates the nodes of a trained artificial neural network (ANN) models and express the ANN outputs as multiple linear regression equation.}
\usage{
pwlnn(Inputdata, ANNOutput, ActualOutput, hiddenW, outputW,
  actfun = "sigmoid", l = NULL, maxBP = NULL, error = NULL,
  sampling = FALSE, samplesize = NULL, ...)
}
\arguments{
\item{Inputdata}{The input data used to trained the ANN model.}

\item{ANNOutput}{The output predicted by the ANN model.}

\item{ActualOutput}{The target output used to train the ANN model.}

\item{hiddenW}{The weights of the hidden neuron in matrix form, columns are the hidden nodes, rows are each of the weights to the hidden node. Only one hidden layer is supported.}

\item{outputW}{The output weights of the output layer. The output activation function is assumed to be linear.}

\item{actfun}{The activation function of the hidden layer. It can be either "sigmoid" or "hyperbolic". Default function is "sigmoid".}

\item{l}{The minimum distance between breakpoints. The minimum value accepted is 1, which is also the default value.}

\item{maxBP}{The maximum number of breakpoints allowed per node. The default value is 6.}

\item{sampling}{TRUE if a sampling subset is used for the PWL approximation. Default value is TRUE.}

\item{Error}{The accepted error rate in terms of percentage difference in MSE with the ANN model. The default value is 20%.}

\item{samplingsize}{The subset size for sampling. Default value is 300.}
}
\value{
pwlnn model The pwl equations of each node, and the predicted output and residuals of the PWL-ANN model.
}
\description{
pwlnn - piecewise linear neural network.
A function that approximates the nodes of a trained artificial neural network (ANN) models and express the ANN outputs as multiple linear regression equation.
}

