% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/takesample.R
\name{takesample}
\alias{takesample}
\title{A function that takes a sample of the given dataset}
\usage{
takesample(data, samplesize)
}
\arguments{
\item{data}{The data that is being sampled}

\item{samplesize}{The desired sample size}
}
\value{
A subset with the specified sample size of the given dataset
}
\description{
A function that takes a sample of the given dataset
}

